test: re-organize mock data and utils
