+++ src/tasks/pair-liquidity-info-history-importer/pair-liquidity-info-history-importer-v2.service.spec.ts
@@ -1,8 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 
-import { Contract } from '@/clients/mdw-http-client.model';
 import { MdwHttpClientService } from '@/clients/mdw-http-client.service';
-import { ContractAddress } from '@/clients/sdk-client.model';
 import { SdkClientService } from '@/clients/sdk-client.service';
 import { PairDbService } from '@/database/pair/pair-db.service';
 import { PairLiquidityInfoHistoryV2DbService } from '@/database/pair-liquidity-info-history/pair-liquidity-info-history-v2-db.service';
@@ -13,117 +11,21 @@ import {
   PairLiquidityInfoHistoryImporterV2Service,
 } from '@/tasks/pair-liquidity-info-history-importer/pair-liquidity-info-history-importer-v2.service';
 import resetAllMocks = jest.resetAllMocks;
+import {
+  contractLog1,
+  contractLog2,
+  contractLog3,
+  contractLog4,
+  contractLog5,
+  contractLog6,
+  contractLog7,
+  contractLog8,
+  initialMicroBlock,
+  pairContract,
+  pairWithTokens,
+} from '@/test/mock-data/pair-liquidity-info-history-mock-data';
 
-// Mock data
-const pair = {
-  id: 1,
-  address: 'ct_pair' as ContractAddress,
-  token0: { address: 'ct_token0' },
-  token1: { address: 'ct_token1' },
-};
-const pairContract: Contract = {
-  aexn_type: '',
-  block_hash: 'mh_hash0',
-  contract: pair.address,
-  source_tx_hash: 'th_',
-  source_tx_type: '',
-  create_tx: {},
-};
-const initialMicroBlock = {
-  hash: pairContract.block_hash,
-  height: '10000',
-  time: '1000000000000',
-};
-const contractLog1 = {
-  args: ['100', '100'],
-  block_hash: 'mh_hash1',
-  block_time: '1000000000001',
-  call_tx_hash: 'th_hash1',
-  call_txi: '10000001',
-  data: '',
-  event_hash: '6O232NLB36RGK54HEJPVDFJVCSIVFV29KPORC07CSSDARM7LV4L0====', // Sync
-  height: '10001',
-  log_idx: '1',
-};
-const contractLog2 = {
-  args: ['123', '100', '100'],
-  block_hash: 'mh_hash1',
-  block_time: '1000000000001',
-  call_tx_hash: 'th_hash1',
-  call_txi: '10000001',
-  data: '',
-  event_hash: 'L2BEDU7I5T8OSEUPB61900P8FJR637OE4MC4A9875C390RMQHSN0====', // PairMint
-  height: '10001',
-  log_idx: '2',
-};
-const contractLog3 = {
-  args: ['200', '200'],
-  block_hash: 'mh_hash2',
-  block_time: '2000000000002',
-  call_tx_hash: 'th_hash2',
-  call_txi: '20000002',
-  data: '',
-  event_hash: '6O232NLB36RGK54HEJPVDFJVCSIVFV29KPORC07CSSDARM7LV4L0====', // Sync
-  height: '20002',
-  log_idx: '1',
-};
-const contractLog4 = {
-  args: ['201', '199'],
-  block_hash: 'mh_hash3',
-  block_time: '3000000000003',
-  call_tx_hash: 'th_hash3',
-  call_txi: '30000003',
-  data: '',
-  event_hash: '6O232NLB36RGK54HEJPVDFJVCSIVFV29KPORC07CSSDARM7LV4L0====', // Sync
-  height: '30003',
-  log_idx: '1',
-};
-const contractLog5 = {
-  args: [],
-  block_hash: 'mh_hash3',
-  block_time: '3000000000003',
-  call_tx_hash: 'th_hash3',
-  call_txi: '30000003',
-  data: '1|0|0|1',
-  event_hash: 'K39AB2I57LEUOUQ04LTEOMSJPJC3G9VGFRKVNJ5QLRMVCMDOPIMG====', // SwapTokens
-  height: '30003',
-  log_idx: '2',
-};
-const contractLog6 = {
-  args: ['100', '100'],
-  block_hash: 'mh_hash3',
-  block_time: '3000000000003',
-  call_tx_hash: 'th_hash4',
-  call_txi: '40000004',
-  data: '',
-  event_hash: '6O232NLB36RGK54HEJPVDFJVCSIVFV29KPORC07CSSDARM7LV4L0====', // Sync
-  height: '30003',
-  log_idx: '1',
-};
-const contractLog7 = {
-  args: [],
-  block_hash: 'mh_hash3',
-  block_time: '3000000000003',
-  call_tx_hash: 'th_hash4',
-  call_txi: '40000004',
-  data: '101|99',
-  event_hash: 'OIS2ALGSJ03MTP2BR5RBFL1GOUGESRVPGE58LGM0MVG9K3VAFKUG====', // PairBurn
-  height: '30003',
-  log_idx: '2',
-};
-const contractLog8 = {
-  args: [],
-  block_hash: 'mh_hash3',
-  block_time: '3000000000003',
-  call_tx_hash: 'th_hash4',
-  call_txi: '40000004',
-  data: '',
-  event_hash: 'non_relevant_event_hash', // Something else
-  height: '30003',
-  log_idx: '3',
-};
-
-const mockMdwClientService = {
+const mockMdwClient = {
   getContract: jest.fn(),
   getMicroBlock: jest.fn(),
   getContractLogsUntilCondition: jest.fn(),
@@ -131,12 +33,12 @@ const mockMdwClientService = {
 
 const mockPairDb = { getAll: jest.fn() };
 
-const mockPairLiquidityInfoHistoryDb = {
+const mockPairLiquidityInfoHistoryV2Db = {
   getLastlySyncedLogByPairId: jest.fn(),
   upsert: jest.fn(),
 };
 
-const mockPairLiquidityInfoHistoryErrorDb = {
+const mockPairLiquidityInfoHistoryV2ErrorDb = {
   getErrorWithinHours: jest.fn(),
   upsert: jest.fn(),
 };
@@ -148,15 +50,15 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
         PairLiquidityInfoHistoryImporterV2Service,
-        { provide: MdwHttpClientService, useValue: mockMdwClientService },
+        { provide: MdwHttpClientService, useValue: mockMdwClient },
         { provide: PairDbService, useValue: mockPairDb },
         {
           provide: PairLiquidityInfoHistoryV2DbService,
-          useValue: mockPairLiquidityInfoHistoryDb,
+          useValue: mockPairLiquidityInfoHistoryV2Db,
         },
         {
           provide: PairLiquidityInfoHistoryV2ErrorDbService,
-          useValue: mockPairLiquidityInfoHistoryErrorDb,
+          useValue: mockPairLiquidityInfoHistoryV2ErrorDb,
         },
         SdkClientService,
       ],
@@ -170,20 +72,20 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
   describe('import', () => {
     it('should import liquidity correctly', async () => {
       // Mock functions
-      mockPairDb.getAll.mockResolvedValue([pair]);
-      mockPairLiquidityInfoHistoryErrorDb.getErrorWithinHours.mockResolvedValue(
+      mockPairDb.getAll.mockResolvedValue([pairWithTokens]);
+      mockPairLiquidityInfoHistoryV2ErrorDb.getErrorWithinHours.mockResolvedValue(
         undefined,
       );
-      mockPairLiquidityInfoHistoryDb.getLastlySyncedLogByPairId
+      mockPairLiquidityInfoHistoryV2Db.getLastlySyncedLogByPairId
         .mockResolvedValueOnce(undefined)
         .mockResolvedValueOnce({
           reserve0: bigIntToDecimal(100n),
           reserve1: bigIntToDecimal(100n),
         });
-      mockMdwClientService.getContract.mockResolvedValue(pairContract);
-      mockMdwClientService.getMicroBlock.mockResolvedValue(initialMicroBlock);
-      mockPairLiquidityInfoHistoryDb.upsert.mockResolvedValue(null);
-      mockMdwClientService.getContractLogsUntilCondition.mockResolvedValue([
+      mockMdwClient.getContract.mockResolvedValue(pairContract);
+      mockMdwClient.getMicroBlock.mockResolvedValue(initialMicroBlock);
+      mockPairLiquidityInfoHistoryV2Db.upsert.mockResolvedValue(null);
+      mockMdwClient.getContractLogsUntilCondition.mockResolvedValue([
         contractLog1,
         contractLog2,
         contractLog3,
@@ -202,24 +104,26 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
 
       // Assertions
       expect(
-        mockPairLiquidityInfoHistoryErrorDb.getErrorWithinHours,
+        mockPairLiquidityInfoHistoryV2ErrorDb.getErrorWithinHours,
       ).toHaveBeenCalledTimes(5); // Once for pair and 4 times for each inserted event
 
       expect(logSpy.mock.calls).toEqual([
         ['Started syncing pair liquidity info history.'],
         ['Syncing liquidity info history for 1 pairs.'],
-        [`Inserted initial liquidity for pair ${pair.id} ${pair.address}.`],
         [
-          `Completed sync for pair ${pair.id} ${pair.address}. Synced 4 log(s).`,
+          `Inserted initial liquidity for pair ${pairWithTokens.id} ${pairWithTokens.address}.`,
+        ],
+        [
+          `Completed sync for pair ${pairWithTokens.id} ${pairWithTokens.address}. Synced 4 log(s).`,
         ],
         ['Finished liquidity info history sync for all pairs.'],
       ]);
 
-      expect(mockPairLiquidityInfoHistoryDb.upsert).toHaveBeenCalledTimes(5);
-      expect(mockPairLiquidityInfoHistoryDb.upsert.mock.calls).toEqual([
+      expect(mockPairLiquidityInfoHistoryV2Db.upsert).toHaveBeenCalledTimes(5);
+      expect(mockPairLiquidityInfoHistoryV2Db.upsert.mock.calls).toEqual([
         [
           {
-            pairId: pair.id,
+            pairId: pairWithTokens.id,
             eventType: 'CreatePair',
             reserve0: bigIntToDecimal(0n),
             reserve1: bigIntToDecimal(0n),
@@ -236,7 +140,7 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
         ],
         [
           {
-            pairId: pair.id,
+            pairId: pairWithTokens.id,
             eventType: EventType.PairMint,
             reserve0: bigIntToDecimal(100n),
             reserve1: bigIntToDecimal(100n),
@@ -253,7 +157,7 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
         ],
         [
           {
-            pairId: pair.id,
+            pairId: pairWithTokens.id,
             eventType: EventType.Sync,
             reserve0: bigIntToDecimal(200n),
             reserve1: bigIntToDecimal(200n),
@@ -270,7 +174,7 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
         ],
         [
           {
-            pairId: pair.id,
+            pairId: pairWithTokens.id,
             eventType: EventType.SwapTokens,
             reserve0: bigIntToDecimal(201n),
             reserve1: bigIntToDecimal(199n),
@@ -287,7 +191,7 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
         ],
         [
           {
-            pairId: pair.id,
+            pairId: pairWithTokens.id,
             eventType: EventType.PairBurn,
             reserve0: bigIntToDecimal(100n),
             reserve1: bigIntToDecimal(100n),
@@ -307,8 +211,8 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
 
     it('should skip a pair if there was a recent error', async () => {
       // Mock functions
-      mockPairDb.getAll.mockResolvedValue([pair]);
-      mockPairLiquidityInfoHistoryErrorDb.getErrorWithinHours.mockResolvedValue(
+      mockPairDb.getAll.mockResolvedValue([pairWithTokens]);
+      mockPairLiquidityInfoHistoryV2ErrorDb.getErrorWithinHours.mockResolvedValue(
         {
           id: 1,
           pairId: 1,
@@ -332,15 +236,15 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
       expect(logSpy.mock.calls).toEqual([
         ['Started syncing pair liquidity info history.'],
         ['Syncing liquidity info history for 1 pairs.'],
-        [`Skipped pair ${pair.id} due to recent error.`],
+        [`Skipped pair ${pairWithTokens.id} due to recent error.`],
         ['Finished liquidity info history sync for all pairs.'],
       ]);
     });
 
     it('should skip a log if there was a recent error', async () => {
       // Mock functions
-      mockPairDb.getAll.mockResolvedValue([pair]);
-      mockPairLiquidityInfoHistoryErrorDb.getErrorWithinHours
+      mockPairDb.getAll.mockResolvedValue([pairWithTokens]);
+      mockPairLiquidityInfoHistoryV2ErrorDb.getErrorWithinHours
         .mockResolvedValueOnce(undefined)
         .mockResolvedValueOnce({
           id: 1,
@@ -354,11 +258,11 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
           updatedAt: Date.now(),
         })
         .mockResolvedValueOnce(undefined);
-      mockPairLiquidityInfoHistoryDb.getLastlySyncedLogByPairId.mockResolvedValue(
+      mockPairLiquidityInfoHistoryV2Db.getLastlySyncedLogByPairId.mockResolvedValue(
         {},
       );
-      mockPairLiquidityInfoHistoryDb.upsert.mockResolvedValue(null);
-      mockMdwClientService.getContractLogsUntilCondition.mockResolvedValue([
+      mockPairLiquidityInfoHistoryV2Db.upsert.mockResolvedValue(null);
+      mockMdwClient.getContractLogsUntilCondition.mockResolvedValue([
         contractLog1,
         contractLog2,
         contractLog4,
@@ -373,7 +277,7 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
 
       // Assertions
       expect(
-        mockPairLiquidityInfoHistoryErrorDb.getErrorWithinHours,
+        mockPairLiquidityInfoHistoryV2ErrorDb.getErrorWithinHours,
       ).toHaveBeenCalledTimes(3); // Once for pair and 2 times for each event
 
       expect(logSpy.mock.calls).toEqual([
@@ -383,15 +287,15 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
           `Skipped log with block hash ${contractLog1.block_hash} tx hash ${contractLog1.call_tx_hash} and log index ${contractLog1.log_idx} due to recent error.`,
         ],
         [
-          `Completed sync for pair ${pair.id} ${pair.address}. Synced 1 log(s).`,
+          `Completed sync for pair ${pairWithTokens.id} ${pairWithTokens.address}. Synced 1 log(s).`,
         ],
         ['Finished liquidity info history sync for all pairs.'],
       ]);
 
-      expect(mockPairLiquidityInfoHistoryDb.upsert.mock.calls).toEqual([
+      expect(mockPairLiquidityInfoHistoryV2Db.upsert.mock.calls).toEqual([
         [
           {
-            pairId: pair.id,
+            pairId: pairWithTokens.id,
             eventType: EventType.SwapTokens,
             reserve0: bigIntToDecimal(201n),
             reserve1: bigIntToDecimal(199n),
@@ -411,7 +315,7 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
 
     it('should catch and insert an error on pair level', async () => {
       const error = {
-        pairId: pair.id,
+        pairId: pairWithTokens.id,
         microBlockHash: '',
         transactionHash: '',
         logIndex: -1,
@@ -419,17 +323,17 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
       };
 
       // Mock functions
-      mockPairDb.getAll.mockResolvedValue([pair]);
-      mockPairLiquidityInfoHistoryErrorDb.getErrorWithinHours.mockResolvedValue(
+      mockPairDb.getAll.mockResolvedValue([pairWithTokens]);
+      mockPairLiquidityInfoHistoryV2ErrorDb.getErrorWithinHours.mockResolvedValue(
         undefined,
       );
-      mockPairLiquidityInfoHistoryDb.getLastlySyncedLogByPairId.mockResolvedValue(
+      mockPairLiquidityInfoHistoryV2Db.getLastlySyncedLogByPairId.mockResolvedValue(
         {},
       );
-      mockMdwClientService.getContractLogsUntilCondition.mockRejectedValue(
+      mockMdwClient.getContractLogsUntilCondition.mockRejectedValue(
         new Error('error'),
       );
-      mockPairLiquidityInfoHistoryErrorDb.upsert.mockResolvedValue(null);
+      mockPairLiquidityInfoHistoryV2ErrorDb.upsert.mockResolvedValue(null);
 
       // Spies
       const logSpy = jest.spyOn(service.logger, 'log');
@@ -439,7 +343,7 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
       await service.import();
 
       // Assertions
-      expect(mockPairLiquidityInfoHistoryErrorDb.upsert).toHaveBeenCalledWith(
+      expect(mockPairLiquidityInfoHistoryV2ErrorDb.upsert).toHaveBeenCalledWith(
         error,
       );
 
@@ -456,7 +360,7 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
 
     it('should catch and insert an error on log level', async () => {
       const error = {
-        pairId: pair.id,
+        pairId: pairWithTokens.id,
         microBlockHash: contractLog3.block_hash,
         transactionHash: contractLog3.call_tx_hash,
         logIndex: parseInt(contractLog3.log_idx),
@@ -464,21 +368,21 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
       };
 
       // Mock functions
-      mockPairDb.getAll.mockResolvedValue([pair]);
-      mockPairLiquidityInfoHistoryErrorDb.getErrorWithinHours.mockResolvedValue(
+      mockPairDb.getAll.mockResolvedValue([pairWithTokens]);
+      mockPairLiquidityInfoHistoryV2ErrorDb.getErrorWithinHours.mockResolvedValue(
         undefined,
       );
-      mockPairLiquidityInfoHistoryDb.getLastlySyncedLogByPairId
+      mockPairLiquidityInfoHistoryV2Db.getLastlySyncedLogByPairId
         .mockResolvedValueOnce({})
         .mockRejectedValueOnce(new Error('error'))
         .mockRejectedValueOnce(undefined);
-      mockMdwClientService.getContractLogsUntilCondition.mockResolvedValue([
+      mockMdwClient.getContractLogsUntilCondition.mockResolvedValue([
         contractLog3,
         contractLog4,
         contractLog5,
       ]);
-      mockPairLiquidityInfoHistoryDb.upsert.mockResolvedValue(null);
-      mockPairLiquidityInfoHistoryErrorDb.upsert.mockResolvedValue(null);
+      mockPairLiquidityInfoHistoryV2Db.upsert.mockResolvedValue(null);
+      mockPairLiquidityInfoHistoryV2ErrorDb.upsert.mockResolvedValue(null);
       const logSpy = jest.spyOn(service.logger, 'log');
       const errorSpy = jest.spyOn(service.logger, 'error');
 
@@ -486,7 +390,7 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
       await service.import();
 
       // Assertions
-      expect(mockPairLiquidityInfoHistoryErrorDb.upsert).toHaveBeenCalledWith(
+      expect(mockPairLiquidityInfoHistoryV2ErrorDb.upsert).toHaveBeenCalledWith(
         error,
       );
 
@@ -494,7 +398,7 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
         ['Started syncing pair liquidity info history.'],
         ['Syncing liquidity info history for 1 pairs.'],
         [
-          `Completed sync for pair ${pair.id} ${pair.address}. Synced 1 log(s).`,
+          `Completed sync for pair ${pairWithTokens.id} ${pairWithTokens.address}. Synced 1 log(s).`,
         ],
         ['Finished liquidity info history sync for all pairs.'],
       ]);
@@ -503,10 +407,10 @@ describe('PairLiquidityInfoHistoryImporterV2Service', () => {
         [`Skipped log. ${JSON.stringify(error)}`],
       ]);
 
-      expect(mockPairLiquidityInfoHistoryDb.upsert.mock.calls).toEqual([
+      expect(mockPairLiquidityInfoHistoryV2Db.upsert.mock.calls).toEqual([
         [
           {
-            pairId: pair.id,
+            pairId: pairWithTokens.id,
             eventType: EventType.SwapTokens,
             reserve0: bigIntToDecimal(201n),
             reserve1: bigIntToDecimal(199n),
