+++ src/tasks/pair-sync/pair-sync.service.spec.ts
@@ -9,14 +9,14 @@ import { PairDbService } from '@/database/pair/pair-db.service';
 import { TokenDbService } from '@/database/token/token-db.service';
 import { Context } from '@/tasks/pair-sync/pair-sync.model';
 import { PairSyncService } from '@/tasks/pair-sync/pair-sync.service';
-import { mockContext } from '@/test/test-utils/context-mock';
-import * as data from '@/test/test-utils/context-mock-data';
-import { mockupEnvVars, TEST_NET_VARS } from '@/test/test-utils/env-mock';
+import { mockContext } from '@/test/utils/context-mock';
+import * as data from '@/test/mock-data/context-mock-data';
+import { mockupEnvVars, TEST_NET_VARS } from '@/test/utils/env-mock';
 import {
   objSubEv,
   swapEvent,
   swapTxInfo,
-} from '@/test/test-utils/subscription-event-mock-data';
+} from '@/test/mock-data/subscription-event-mock-data';
 
 describe('PairSyncService', () => {
   let service: PairSyncService;
