+++ test/e2e/pair-sync.service.e2e-spec.ts
@@ -7,9 +7,9 @@ import { PairDbService } from '@/database/pair/pair-db.service';
 import { PrismaService } from '@/database/prisma.service';
 import { TokenDbService } from '@/database/token/token-db.service';
 import { PairSyncService } from '@/tasks/pair-sync/pair-sync.service';
-import { mockContext } from '@/test/test-utils/context-mock';
-import * as data from '@/test/test-utils/context-mock-data';
-import { cleanDb } from '@/test/test-utils/db-helper';
+import { mockContext } from '@/test/utils/context-mock';
+import * as data from '@/test/mock-data/context-mock-data';
+import { cleanDb } from '@/test/utils/db-helper';
 
 // Testing method
 // 1. before all create a common context
