+++ test/e2e/pairs.e2e-spec.ts
@@ -10,10 +10,10 @@ import { PairDbService } from '@/database/pair/pair-db.service';
 import { PrismaService } from '@/database/prisma.service';
 import { TokenDbService } from '@/database/token/token-db.service';
 import { PairSyncService } from '@/tasks/pair-sync/pair-sync.service';
-import { mockContext } from '@/test/test-utils/context-mock';
-import * as data from '@/test/test-utils/context-mock-data';
-import { cleanDb, listToken } from '@/test/test-utils/db-helper';
-import { sortByAddress } from '@/test/test-utils/utils';
+import { mockContext } from '@/test/utils/context-mock';
+import * as data from '@/test/mock-data/context-mock-data';
+import { cleanDb, listToken } from '@/test/utils/db-helper';
+import { sortByAddress } from '@/test/utils/utils';
 
 // Testing method
 // before all
